+++ nautilus_trader/adapters/interactive_brokers/factories.py
@@ -49,7 +49,9 @@ def get_cached_ib_client(
     host: str = "127.0.0.1",
     port: int = 4001,
     connect=True,
-    timeout=90,
+    timeout=180,
+    client_id: int = 1,
+    start_gateway: bool = True,
 ) -> ib_insync.IB:
     """
     Cache and return a InteractiveBrokers HTTP client with the given key and secret.
@@ -71,6 +73,10 @@ def get_cached_ib_client(
         Whether to connect to IB.
     timeout: int, optional
         The timeout for trying to establish a connection
+    client_id: int, optional
+        The client_id to connect with
+    start_gateway: bool
+        Start the IB Gateway docker container
 
     Returns
     -------
@@ -78,11 +84,11 @@ def get_cached_ib_client(
 
     """
     global IB_INSYNC_CLIENTS, GATEWAY
-
-    # Start gateway
-    if GATEWAY is None:
-        GATEWAY = InteractiveBrokersGateway(username=username, password=password)
-        GATEWAY.safe_start()
+    if start_gateway:
+        # Start gateway
+        if GATEWAY is None:
+            GATEWAY = InteractiveBrokersGateway(username=username, password=password)
+            GATEWAY.safe_start()
 
     client_key: tuple = (host, port)
 
@@ -90,7 +96,7 @@ def get_cached_ib_client(
         client = ib_insync.IB()
         if connect:
             try:
-                client.connect(host=host, port=port, timeout=timeout)
+                client.connect(host=host, port=port, timeout=timeout, clientId=client_id)
             except TimeoutError:
                 raise TimeoutError(f"Failed to connect to gateway in {timeout}s")