+++ nautilus_trader/analysis/analyzer.py
@@ -54,13 +54,8 @@ class PortfolioAnalyzer:
         statistic : PortfolioStatistic
             The statistic to register.
 
-        Raises
-        ------
-        KeyError if `statistic` has already been registered.
-
         """
         PyCondition.not_none(statistic, "statistic")
-        PyCondition.not_in(statistic.name, self._statistics, "statistic.name", "_statistics")
 
         self._statistics[statistic.name] = statistic
 
@@ -150,7 +145,6 @@ class PortfolioAnalyzer:
 
         """
         self._positions += positions
-
         for position in positions:
             self.add_trade(position.id, position.realized_pnl)
             self.add_return(unix_nanos_to_dt(position.ts_closed), position.realized_return)