+++ nautilus_trader/analysis/statistics/returns_volatility.py
@@ -21,14 +21,22 @@ import pandas as pd
 from nautilus_trader.analysis.statistic import PortfolioStatistic
 
 
-class ReturnsAnnualVolatility(PortfolioStatistic):
+class ReturnsVolatility(PortfolioStatistic):
     """
-    Calculates the annualized volatility of returns.
+    Calculates the volatility of returns.
+
+    Parameters
+    ----------
+    period : int, default 252
+        The trading period in days.
     """
+
+    def __init__(self, period: int = 252):
+        self.period = period
 
     @property
     def name(self) -> str:
-        return "Annual Volatility (Returns)"
+        return f"Returns Volatility ({self.period} days)"
 
     def calculate_from_returns(self, returns: pd.Series) -> Optional[Any]:
-        return returns.std() * np.sqrt(252)
+        return returns.std() * np.sqrt(self.period)