+++ nautilus_trader/cache/cache.pyx
@@ -13,6 +13,9 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+import copy
+import pickle
+import uuid
 from collections import deque
 from decimal import Decimal
 from typing import Optional
@@ -101,6 +104,7 @@ cdef class Cache(CacheFacade):
         self._accounts = {}                    # type: dict[AccountId, Account]
         self._orders = {}                      # type: dict[ClientOrderId, Order]
         self._positions = {}                   # type: dict[PositionId, Position]
+        self._position_snapshots = {}          # type: dict[PositionId, list[bytes]]
 
         # Cache index
         self._index_venue_account = {}         # type: dict[Venue, AccountId]
@@ -548,6 +552,7 @@ cdef class Cache(CacheFacade):
         self._accounts.clear()
         self._orders.clear()
         self._positions.clear()
+        self._position_snapshots.clear()
 
         self._log.debug(f"Cleared cache.")
 
@@ -1270,6 +1275,33 @@ cdef class Cache(CacheFacade):
         # Update database
         if self._database is not None:
             self._database.add_position(position)
+
+    cpdef void snapshot_position(self, Position position) except *:
+        """
+        Snapshot the given position in its current state.
+
+        The position ID will be appended with a UUID4 string.
+
+        Parameters
+        ----------
+        position : Position
+            The position to archive.
+
+        """
+        cdef PositionId position_id = position.id
+        cdef list snapshots = self._position_snapshots.get(position_id)
+
+        # Reassign position ID
+        cdef Position copied_position = copy.copy(position)
+        copied_position.id = PositionId(position.id.value + f"-{uuid.uuid4()}")
+        cdef bytes position_pickled = pickle.dumps(copied_position)
+
+        if snapshots is not None:
+            snapshots.append(position_pickled)
+        else:
+            self._position_snapshots[position_id] = [position_pickled]
+
+        self._log.debug(f"Snapshot {repr(copied_position)}.")
 
     cpdef void update_account(self, Account account) except *:
         """
@@ -2518,6 +2550,32 @@ cdef class Cache(CacheFacade):
         Condition.not_none(client_order_id, "client_order_id")
 
         return self._index_order_position.get(client_order_id)
+
+    cpdef list position_snapshots(self, PositionId position_id=None):
+        """
+        Return all position snapshots with the given optional identifier filter.
+
+        Parameters
+        ----------
+        position_id : PositionId, optional
+            The position ID query filter.
+
+        Returns
+        -------
+        list[Position]
+
+        """
+        cdef list snapshot_list
+        cdef list snapshots
+        if position_id is not None:
+            snapshots = self._position_snapshots.get(position_id, [])
+        else:
+            snapshots = []
+            for snapshot_list in self._position_snapshots.values():
+                snapshots += snapshot_list
+
+        cdef bytes s
+        return [pickle.loads(s) for s in snapshots]
 
     cpdef list positions(
         self,