+++ tests/unit_tests/cache/test_cache_execution.py
@@ -373,6 +373,40 @@ class TestCache:
         assert self.cache.position_for_order(order.client_order_id) == position
         assert self.cache.orders_for_position(position.id) == [order]
 
+    def test_snapshot_position(self):
+        # Arrange
+        order = self.strategy.order_factory.market(
+            AUDUSD_SIM.id,
+            OrderSide.BUY,
+            Quantity.from_int(100000),
+        )
+
+        position_id = PositionId("P-1")
+        self.cache.add_order(order, position_id)
+
+        fill = TestEventStubs.order_filled(
+            order,
+            instrument=AUDUSD_SIM,
+            position_id=PositionId("P-1"),
+            last_px=Price.from_str("1.00000"),
+        )
+
+        position = Position(instrument=AUDUSD_SIM, fill=fill)
+
+        # Act
+        self.cache.snapshot_position(position)
+        self.cache.snapshot_position(position)
+        snapshots = self.cache.position_snapshots(position.id)
+
+        # Assert
+        assert len(snapshots) == 2
+        assert snapshots[0].id.value.startswith(position.id.value)
+        snapshot_dict = snapshots[0].to_dict()
+        del snapshot_dict["position_id"]
+        position_dict = position.to_dict()
+        del position_dict["position_id"]
+        assert snapshot_dict == position_dict
+
     def test_load_position(self):
         # Arrange
         order = self.strategy.order_factory.market(